requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
instaloader==4.11
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
